LayoutTests/js/dom/modules/nomodule-prevents-execution-classic-script-src.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test nomodule prevents execution of classic script.');
window.executed = false;
</script>
<script nomodule src="./script-tests/module-src-simple.js"></script>
<script>
shouldBeFalse(`executed`);
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
